refactor:replace all files into the correct places
