+++ .gitignore
@@ -7,8 +7,7 @@ __pycache__/
 *.so
 
 # Notebook files
-notebooks/*.ipynb
-*.ipynb
+notebooks
 
 # Distribution / packaging
 .Python
